LIBAVALON-182. Add /sitemap.py
This commit represents an initial configuration to use the Avalon API to
build the sitemap. However, this approach will be abandoned due to
the very slow response times using the Avalon API to pull media
object information, likely because Avalon is performaning a full fetch
from Fedora for each request, resulting in a rate of 2 objects / second.
Once Avalon is fully populated this could lead to build times of
1-2 hours.

TODO: Build the sitemap using Solr

https://issues.umd.edu/browse/LIBAVALON-182